backend/requirements.txt
# ================================
# AccountIA Backend Requirements (MVP Mínimo)
# ================================

# Django Core
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
firebase-admin==6.2.0
PyJWT==2.8.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.27.2

# Task Queue (Celery) - Básico
celery==5.3.4
redis==5.0.1

# File Processing
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.0.1
python-dotenv==1.0.0
numpy==1.24.4
xlrd==2.0.1

# Google Cloud (Básico)
google-cloud-storage==2.10.0
google-generativeai==0.3.2

# Utilities
requests==2.31.0
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3

# Development Tools (Básico)
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0